Time series anomaly detection/requirements.txt
# requirements.txt
# Core data science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
joblib>=1.1.0

# Deep learning
tensorflow>=2.8.0
# or pytorch if preferred:
# torch>=1.10.0
# torchvision>=0.11.0

# Time series analysis
statsmodels>=0.13.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0

# Configuration and logging
pyyaml>=6.0
tqdm>=4.62.0

# Utilities
python-dotenv>=0.19.0
ipykernel>=6.0.0
jupyter>=1.0.0

# Development
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0